+++ packages/explorerkit-idls/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @solanafm/explorer-kit-idls
 
+## 1.0.3
+
+### Patch Changes
+
+- 3373713: fix: fix new extensions not able to serialize new extensions in mint account"
+
+## 1.0.2
+
+### Patch Changes
+
+- f18f2f3: feat: token extensions full supoprt till v1
+
 ## 1.0.1
 
 ### Patch Changes